+++ Runtime/Main/AccelByteSDKMain.cs
@@ -1,0 +1,51 @@
+// Copyright (c) 2023 AccelByte Inc. All Rights Reserved.
+// This is licensed software from AccelByte Inc, for limitations
+// and restrictions contact your company contract manager.
+
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace AccelByte.Core
+{
+#if (UNITY_STANDALONE_LINUX || UNITY_EDITOR_LINUX) && UNITY_SERVER
+    using PlatformMain = LinuxServerMain;
+#else
+    using PlatformMain = NullMain;
+#endif
+    internal static class AccelByteSDKMain
+    {
+        internal static PlatformMain Main;
+        
+        [RuntimeInitializeOnLoadMethod]
+        private static void StartAccelByteSDK()
+        {
+            if (Main == null)
+            {
+                Main = new PlatformMain();
+            }
+
+            AccelByteDebug.LogVerbose("AccelByte API Initialized!");
+            Main.Run();
+#if UNITY_EDITOR
+            UnityEditor.EditorApplication.playModeStateChanged += EditorApplicationPlayyModeStateChanged;
+#endif
+            Application.quitting += ApplicationQuitting;
+        }
+
+#if UNITY_EDITOR
+        private static void EditorApplicationPlayyModeStateChanged(UnityEditor.PlayModeStateChange newState)
+        {
+            if(newState == UnityEditor.PlayModeStateChange.ExitingPlayMode)
+            {
+                UnityEditor.EditorApplication.playModeStateChanged -= EditorApplicationPlayyModeStateChanged;
+                Main.Stop();
+            }
+        }
+#endif
+        private static void ApplicationQuitting()
+        {
+            Main.Stop();
+        }
+    }
+}