+++ Runtime/Main/IPlatformMain.cs
@@ -1,0 +1,11 @@
+// Copyright (c) 2023 AccelByte Inc. All Rights Reserved.
+// This is licensed software from AccelByte Inc, for limitations
+// and restrictions contact your company contract manager.
+namespace AccelByte.Core
+{
+    internal interface IPlatformMain
+    {
+        public void Run();
+        public void Stop();
+    }
+}